requirements.txt

streamlit>=1.33
opencv-python
numpy
pillow
# Choose a TensorFlow build appropriate to your OS/CPU/GPU
# For most Linux/Windows (CPU):
tensorflow>=2.12
# On Apple Silicon macOS, replace the above line with:
# tensorflow-macos>=2.12
